Added new DB for messaging
