+++ treetracker-core/Persistence/CoreDataManager.swift
@@ -11,37 +11,9 @@ import CoreData
 class CoreDataManager: CoreDataManaging {
 
     private lazy var persistentContainer: NSPersistentContainer = {
-
         let bundle = Bundle(for: CoreDataManager.self)
-
-        guard let modelURL = bundle.url(forResource: "TreeTracker", withExtension: "momd") else {
-            fatalError("Unable to load managed object model URL")
-        }
-
-        guard let managedObjectModel = NSManagedObjectModel(contentsOf: modelURL) else {
-            fatalError("Unable to load managed object model: \(modelURL)")
-        }
-
-        let container = NSPersistentContainer(name: "TreeTracker", managedObjectModel: managedObjectModel)
-        container.loadPersistentStores { _, error in
-            if let error = error {
-                fatalError("Unable to load persistent stores: \(error)")
-            }
-        }
-        return container
+        return makePersistentContainer(modelName: "TreeTracker", bundle: bundle)
     }()
-
-    func saveContext() {
-        let context = persistentContainer.viewContext
-        if context.hasChanges {
-            do {
-                try context.save()
-            } catch {
-                let nserror = error as NSError
-                fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
-            }
-        }
-    }
 
     var viewContext: NSManagedObjectContext {
         return persistentContainer.viewContext
